Refactored folder discovery into a helper
We would like to maintain the existing directory enumeration logic, but we want to add a second directory to enumerate. Rather than futzing with the logic, we refactor it out.

Includes partial work for supporting the new switch.

Additional: Update `launch.json` to move the data deeper to allow one single directory in `.gitignore`